+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 ##
 ## [5.3.0] - tbd
+### New
+ - Add label to public shares
 ### Fixes
  - Fix API calls
 ### Changes