+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 ##
-## [5.3.0] - tbd
+## [5.3.0] - 2023-09-06
 ### New
  - Add label to public shares
  - Send all unsent invitations to a poll with one click (resolves contact groups and circles too)